requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.8
django-bootstrap-v5==1.0.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Pillow==10.1.0
PyJWT==2.8.0
pytz==2023.3.post1
rest-framework-simplejwt==0.0.2
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
